fixing router for signin singnup + move component addcard to the good route
