+++ src/pages/CreateProject/CreateProject.js
@@ -1,9 +1,11 @@
 import React from 'react';
-import Log from "../components/Log";
+import Log from "../../components/Log";
+import NewPostForm from "../../components/Post/NewPostForm/NewPostForm";
 
 const CreateProject = () => {
     return (
         <div>
+            <NewPostForm/>
             page add project
         </div>
     );
